requirements.txt
# --- Core runtime ---
python-dotenv>=1.0.1
pydantic>=2.7.0
numpy>=1.24
requests>=2.32.0

# --- Streamlit app ---
# Core
streamlit==1.38.0

# Preferred mic widget (if it builds for you)
streamlit-mic-recorder==0.0.4

# Fallback mic widget
audio-recorder-streamlit>=0.0.8
watchdog>=4.0.1

# --- OpenAI + ElevenLabs (optional AI APIs) ---
openai>=1.40.0
langchain>=0.2.15
langchain-openai>=0.1.23
elevenlabs>=1.3.0
aiohttp>=3.9.5

# --- Resume parsing ---
docx2txt>=0.8

pypdf>=4.3.1           # fallback if PyMuPDF unavailable

# --- Light STT (optional fallback) ---
faster-whisper>=1.0.3    # safe pure-Python option; optional

# --- Feedback scoring / ML utils ---
scikit-learn>=1.5.0


# --- Optional database ---
pymongo>=4.8.0

# --- Dev / linting helpers ---
black>=24.4.2
ruff>=0.5.5

faiss-cpu
sentence-transformers
tiktoken                  # optional (OpenAI embeddings path)
